Revert "fix(subsonic): duration should be integer"
This reverts commit 018ecebf2cc6bce451620e581730e3f4f9688a29.

+++ src/models/songs.rs
@@ -28,7 +28,7 @@ pub struct SongUpdateInformationDB<'a> {
     pub languages: Vec<&'static str>,
     // Song property
     pub format: Cow<'a, str>,
-    pub duration: i32,
+    pub duration: f32,
     pub bitrate: i32,
     pub sample_rate: i32,
     pub channel_count: i16,
@@ -74,7 +74,7 @@ pub mod test {
         pub original_release_day: Option<i16>,
         pub languages: Vec<Option<String>>,
         // Song property
-        pub duration: i32,
+        pub duration: f32,
         // Filesystem property
         pub music_folder_id: Uuid,
         pub relative_path: String,
